requirements.txt
telethon
asyncio
#sqlite3
python-dotenv
logging
